requirements.txt
Django==2.2.8
djangorestframework==3.10.3
django-cors-headers==2.4.0
django-filter==2.2.0
djangorestframework_simplejwt
